Exercises 4.21

+++ Part-4/controllers/blogs.js
@@ -53,6 +53,28 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  let decodedToken
+
+  try {
+    decodedToken = jwt.verify(request.token, config.SECRET)
+  } catch (error) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'only the creator can delete a blog' })
+  }
+
   const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
 
   if (deletedBlog?.user) {
